wiring for ExtractorRoutingService
