+++ app/services/mongodb/MongoDBFileService.scala
@@ -62,6 +62,7 @@ class MongoDBFileService @Inject() (
   folders: FolderService,
   metadatas: MetadataService,
   events: EventService,
+  routing: ExtractorRoutingService,
   appConfig: AppConfigurationService,
   esqueue: ElasticsearchQueue) extends FileService {
 
@@ -190,15 +191,9 @@ class MongoDBFileService @Inject() (
       datasetId = datasetslists.head.id
     }
     val extractorId = play.Play.application().configuration().getString("archiveExtractorId")
-    val plugins = current.plugin[RabbitmqPlugin]
-    plugins.foreach { p =>
-      p.submitFileManually(new UUID(originalId), file, host, extractorId, extra,
-        datasetId, newFlags, apiKey, user)
-      Logger.info("Sent archive request for file " + id)
-    }
-    if (plugins.isEmpty) {
-      Logger.warn("RabbitMQ plugin not detected: archival may be disabled")
-    }
+    routing.submitFileManually(new UUID(originalId), file, host, extractorId, extra,
+      datasetId, newFlags, apiKey, user)
+    Logger.info("Sent archive request for file " + id)
   }
 
   /**
